Extracted all base auth functionality out into the auth-core module.
