+++ keycloak4s-auth-core/src/main/scala/com/fullfacing/keycloak4s/auth/core/PolicyBuilders.scala
@@ -1,6 +1,6 @@
-package com.fullfacing.keycloak4s.auth.akka.http.authorization
+package com.fullfacing.keycloak4s.auth.core
 
-import com.fullfacing.keycloak4s.auth.core.Logging
+import com.fullfacing.keycloak4s.auth.core.authorization.PathAuthorization
 import com.fullfacing.keycloak4s.core.Exceptions
 
 import scala.io.{BufferedSource, Source}
