--- keycloak4s-auth/akka-http/src/main/scala/com/fullfacing/keycloak4s/auth/akka/http/directives/Directives.scala
@@ -1,3 +1,0 @@
-package com.fullfacing.keycloak4s.auth.akka.http.directives
-
-object Directives extends ValidationDirective with SecurityDirectives
