+++ keycloak4s-auth/akka-http/src/main/scala/com/fullfacing/keycloak4s/auth/akka/http/directives/SecurityDirective.scala
@@ -4,7 +4,7 @@ import akka.http.scaladsl.server.Directive1
 import com.fullfacing.keycloak4s.auth.akka.http.directives.magnets.SecurityMagnet
 import com.fullfacing.keycloak4s.auth.core.models.AuthPayload
 
-trait SecurityDirectives extends ValidationDirective {
+object SecurityDirective {
 
   /**
    * Authorization directive that secures all inner routes.
