+++ keycloak4s-auth/akka-http/src/main/scala/com/fullfacing/keycloak4s/auth/akka/http/directives/ValidationDirective.scala
@@ -11,7 +11,7 @@ import com.fullfacing.keycloak4s.core.models.KeycloakException
 
 import scala.util.{Failure, Success}
 
-trait ValidationDirective {
+object ValidationDirective {
 
   type AuthPayloadWithId = (UUID, AuthPayload)
