+++ keycloak4s-playground/src/test/scala/AuthorizationTests.scala
@@ -3,7 +3,8 @@ import java.util.concurrent.atomic.AtomicReference
 
 import akka.http.scaladsl.model.HttpMethods
 import akka.http.scaladsl.model.Uri.Path
-import com.fullfacing.keycloak4s.auth.akka.http.authorization.{PathAuthorization, PolicyBuilders}
+import com.fullfacing.keycloak4s.auth.core.PolicyBuilders
+import com.fullfacing.keycloak4s.auth.core.authorization.PathAuthorization
 import com.fullfacing.keycloak4s.core.models.enums.PolicyEnforcementModes
 import org.scalatest.{FlatSpec, Matchers}
 import utils.AuthTestData
@@ -42,7 +43,7 @@ class AuthorizationTests extends FlatSpec with Matchers {
     val method    = HttpMethods.GET
     val userRoles = List("read")
 
-    val result = config.get().authorizeRequest(path, method, userRoles)
+    val result = config.get().authorizeRequest(path.toString(), method.value, userRoles)
 
     result shouldBe true
   }
@@ -52,7 +53,7 @@ class AuthorizationTests extends FlatSpec with Matchers {
     val method    = HttpMethods.HEAD
     val userRoles = List("wildcard-role")
 
-    val result = config.get().authorizeRequest(path, method, userRoles)
+    val result = config.get().authorizeRequest(path.toString(), method.value, userRoles)
 
     result shouldBe true
   }
@@ -62,7 +63,7 @@ class AuthorizationTests extends FlatSpec with Matchers {
     val method    = HttpMethods.HEAD
     val userRoles = List("delete")
 
-    val result = config.get().authorizeRequest(path, method, userRoles)
+    val result = config.get().authorizeRequest(path.toString(), method.value, userRoles)
 
     result shouldBe false
   }
@@ -72,7 +73,7 @@ class AuthorizationTests extends FlatSpec with Matchers {
     val method    = HttpMethods.HEAD
     val userRoles = List("wildcard-role")
 
-    val result = config.get().authorizeRequest(path, method, userRoles)
+    val result = config.get().authorizeRequest(path.toString(), method.value, userRoles)
 
     result shouldBe false
   }
@@ -82,7 +83,7 @@ class AuthorizationTests extends FlatSpec with Matchers {
     val method    = HttpMethods.PATCH
     val userRoles = List("read")
 
-    val result = config.get().authorizeRequest(path, method, userRoles)
+    val result = config.get().authorizeRequest(path.toString(), method.value, userRoles)
 
     result shouldBe false
   }
@@ -92,7 +93,7 @@ class AuthorizationTests extends FlatSpec with Matchers {
     val method    = HttpMethods.GET
     val userRoles = List("read")
 
-    val result = config.get().authorizeRequest(path, method, userRoles)
+    val result = config.get().authorizeRequest(path.toString(), method.value, userRoles)
 
     result shouldBe false
   }
@@ -102,7 +103,7 @@ class AuthorizationTests extends FlatSpec with Matchers {
     val method    = HttpMethods.DELETE
     val userRoles = List("read", "write")
 
-    val result = config.get().authorizeRequest(path, method, userRoles)
+    val result = config.get().authorizeRequest(path.toString(), method.value, userRoles)
 
     result shouldBe false
   }
@@ -112,7 +113,7 @@ class AuthorizationTests extends FlatSpec with Matchers {
     val method    = HttpMethods.GET
     val userRoles = List("no")
 
-    val result = config.get().authorizeRequest(path, method, userRoles)
+    val result = config.get().authorizeRequest(path.toString(), method.value, userRoles)
 
     result shouldBe false
   }
